complete-loginAndRegister-form

+++ index.php
@@ -1,17 +1,29 @@
 <?php
 include "bootstrap/init.php";
 
-if(isset($_GET["delete_folder"]) && is_numeric($_GET["delete_folder"])){
-    deleteFolder($_GET["delete_folder"]);
+
+if(isLoggedIn()){
     
-}
+    if(isset($_GET["logout"]) && $_GET['logout'] == "true"){
+        logout();
+        die();
+    }
 
-if(isset($_GET["delete_task"]) && is_numeric($_GET["delete_task"])){
-    deleteTask($_GET["delete_task"]);
+
+    if(isset($_GET["delete_folder"]) && is_numeric($_GET["delete_folder"])){
+        deleteFolder($_GET["delete_folder"]);
+        
+    }
     
+    if(isset($_GET["delete_task"]) && is_numeric($_GET["delete_task"])){
+        deleteTask($_GET["delete_task"]);
+        
+    }
+    $folders = getFolders();
+    $tasks = getTasks();
+    
+    
+    include "template/tpl-index.php";
+}else {
+    redirect(site_url("auth.php"));
 }
-$folders = getFolders();
-$tasks = getTasks();
-
-
-include "template/tpl-index.php";
